Move templiert's implementation of class SEM_MultiSEM to sem_control_zeiss.py
